cybersecurity-containerlabs/command-control/dns_tunneling/server/requirements.txt
dnslib==0.9.24
python-dotenv==1.0.1
